added category sub-category and child category
