+++ app/Http/Controllers/CategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Category;
+use App\Models\Gallery;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\URL;
 use Illuminate\Support\Facades\Request;
@@ -17,19 +18,22 @@ class CategoryController extends Controller
      */
     public function index()
     {
-        return inertia('Category', [
-            $search = Request::input('search'),
-            Category::query()->when($search, function ($query, $search) {
+
+
+        return inertia('Backend/Category', [
+            'categories' => Category::query()->with("photos")
+                ->when(Request::input('search'), function ($query, $search) {
                     $query->where('name', 'like', "%{$search}%");
                 })
                 ->paginate(Request::input('perPage') ?? 10)
                 ->withQueryString()
                 ->through(fn($category) => [
-                    'id' => $category->id,
-                    'name' => $category->name,
-                    'slug' => $category->slug,
+                    'id'    => $category->id,
+                    'name'  => $category->name,
+                    'slug'  => $category->slug,
+                    'pic' => $category,
+                    'photo' => global_asset($category->photos->count() > 0 ? $category->photos[0]->filename : null)
                 ]),
-            'parent_categories' => Category::where('parent_id', 0)->get(),
             'filters' => Request::only(['search','perPage']),
             'url' => URL::route('categories.index')
         ]);
@@ -53,12 +57,19 @@ class CategoryController extends Controller
      */
     public function store()
     {
-        Category::create(
-            Request::validate([
-                'name' => 'required|max:50',
-                'photo' => 'required',
-            ])
-        );
+        Request::validate([
+            'name' => 'required|max:50',
+            'photo' => 'required',
+        ]);
+        $cat = Category::create(['name' => Request::input('name')]);
+        if(Request::hasFile('photo')){
+            $path = Request::file('photo')->store('image', 'public');
+        }
+        Gallery::create([
+            'imageable_id' => $cat->id,
+            'imageable_type' => 'App\\Models\\Category',
+            'filename' => $path
+        ]);
 
         return Redirect::route('categories.index');
     }
