+++ resources/js/components/NotfoundSvg.vue
@@ -1,0 +1,14 @@
+<template>
+
+    
+</template>
+
+<script>
+export default {
+    name: "NotfoundSvg"
+}
+</script>
+
+<style scoped>
+
+</style>
